added shit animation + collider wall
Added a bad animation just to test animations out. It is not the character we are going to use so does not really matter. Besides that I added a tilemap collider to the tilemap for the walls. It is quite buggy at some edges, needs a bit of debugging still.