added a gun + fixed some stuff
Added a weapon that deals 10-25 (random) damage to enemies on hit. Right now it is still an ugly cube but the bullet animations are sweet :)

Fixed some camera issues.

Also started experimenting with enemies freezing in the flashlight (is broken as of right now, so one enemy is not working properly)