Final balance changes nurse room
I think all rooms are great now, nurse is difficulit to figure out, but after that its a matter of having a good tactic/walking path. Boss room is difficult, but possible, which is great. Darkness room also hard to figure out, but after that gets more and more doable.